+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 ### Compatible changes
+* Switch to tightvncserver to be compatible with Ubuntu 20.04. vnc4server is still supported.
+
 ### Breaking changes